requirements.txt
aiohttp
asyncio-gevent
#audioop-lts #uncomment for 3.13+
Flask
flask-sock
gevent
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
MarkupSafe
PyAudio
pynput
python-vlc
requests
twitchio
twitchio[starlette]
websocket-client
